encryption key is no longer hardcoded

+++ hardware/victims/firmware/crypto/galois-aes128/aes.h
@@ -10,7 +10,7 @@
 #define AES_CONST_VAR
 #endif
 
-void GALOIS_AES128_setkey(uint32_t *key);
+void GALOIS_AES128_setkey(uint8_t* key);
 void GALOIS_AES128_blockencrypt(uint8_t* pt);
 
 #endif //_AES_H_
